pyproject.toml
[project]
name = "dinomite"
version = "0.1.0"
description = "DINOmite"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "joblib>=1.5.1",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.5",
    "pandas>=2.0.0",
    "pillow>=10.0.0",
    "scikit-learn>=1.7.1",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers",
    "autoattack @ git+https://github.com/fra31/auto-attack",
]

[dependency-groups]
dev = ["pre-commit>=4.3.0", "ruff>=0.12.8"]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", marker = "sys_platform == 'darwin'" },
    { index = "pytorch-cuda", marker = "sys_platform != 'darwin'" },
]
torchvision = [
    { index = "pytorch-cpu", marker = "sys_platform == 'darwin'" },
    { index = "pytorch-cuda", marker = "sys_platform != 'darwin'" },
]
transformers = { git = "https://github.com/huggingface/transformers.git" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cuda"
url = "https://download.pytorch.org/whl/cu129"
explicit = true

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "double"
